sunstrike_component/sunstrike_component/pyproject.toml
[tool.poetry]
name = "sunstrike-component"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.2.1"
pulsar-client = "^3.5.0"

[tool.setuptools]
packages = ["sunstrike_component"]

[build-system]
requires = [
    "poetry-core"
]
build-backend = "poetry.core.masonry.api"
